+++ src/components/home/MemeListContainer/MemeList/index.ts
@@ -1,3 +1,2 @@
 export * from "./CommonMemeList";
-export * from "./InfiniteMemeList";
 export * from "./UserFindMemeList";
